remove useless features
